requirements.txt

aiocron==1.8
anyio==3.6.2
APScheduler==3.6.3
blinker==1.6.2
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
croniter==1.3.14
dnspython==2.3.0
fastapi==0.95.1
Flask==2.3.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2023.3
requests==2.30.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tornado==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
uvicorn==0.22.0
Werkzeug==2.3.4
